deployment/sample-agents/azure-research-agent/requirements.txt
azure-identity>=1.15.0
openai>=1.0.0
azure-search-documents>=11.4.0
azure-core>=1.29.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
